refactor(renderer): move AwsmRenderer + impls lib.rs → renderer.rs
lib.rs (crate root) now holds only the crate doc, #![allow] crate attributes,
the pub mod tree, pub mod core re-export, and mod renderer; pub use renderer::*.
All logic (AwsmRenderer struct + builder + impls + free fns + states) moved to
renderer.rs (use crate::* brings the crate-root module names + re-exports into
scope so the original crate-root-relative paths resolve unchanged).

Two necessary, behavior-preserving visibility touch-ups (the struct relocating
off the crate root changes who counts as a 'descendant' for private access;
both keep identical crate-internal visibility + unchanged external API):
- AwsmRenderer._clear_color{,_perceptual_to_linear}: private → pub(crate)
  (read by sibling render.rs).
- crate root re-adds pub(crate) use debug::AwsmRendererLogging (was provided by
  the now-relocated private use-at-root; frontends already import it via the
  debug:: path, so external API is unchanged).

Zero behavior change: cargo check + clippy clean, 160 tests pass, dependent
libs (gltf, editor) compile.
